+++ riscv-ovpsim/source/riscvStructure.h
@@ -61,7 +61,9 @@
 //
 // Number of temporaries
 //
-#define NUM_TEMPS 4
+#define NUM_TEMPS_WIDE  3
+#define NUM_TEMPS_BASIC 4
+#define NUM_TEMPS       (NUM_TEMPS_WIDE+NUM_TEMPS_BASIC)
 
 //
 // Container for net values
@@ -126,6 +128,7 @@ typedef struct riscvNetPortS {
 //
 #define VLEN_MAX        2048
 #define VBYTES_MAX      (VLEN_MAX/8)
+#define VDWORDS_MAX     (VLEN_MAX/64)
 #define VREG_NUM        32
 #define ELEN_MIN        32
 #define ELEN_MAX        64
@@ -142,7 +145,7 @@ typedef struct riscvNetPortS {
 // of up to VLEN_MAX bits. The assignment of the storage depends on the
 // configured VLEN
 //
-typedef Uns64 riscvVRegBank[(VLEN_MAX*VREG_NUM)/64];
+typedef Uns64 riscvVRegBank[VDWORDS_MAX*VREG_NUM];
 
 //
 // This defines the type of elements of the stride tables used to handle
@@ -244,12 +247,13 @@ typedef struct riscvS {
 
     // Vector extension
     Uns8               vTypeKey;                    // vector polymorphic configuration
+    Uns8               vFieldMask;                  // vector field mask
+    Uns8               vActiveMask;                 // vector active element mask
     Bool               vFirstFault;                 // vector first fault active?
     Uns32              vlMax;                       // maximum vl element number
     riscvVRegBank      v;                           // vector registers (configurable size)
     UnsPS              vBase[NUM_BASE_REGS];        // indexed base registers
     UnsPS              offsetBase;                  // offset table base register
-    riscvStrideOffset  offsetsLMULx1[VBYTES_MAX*1]; // LMULx1 stride offsets
     riscvStrideOffset  offsetsLMULx2[VBYTES_MAX*2]; // LMULx2 stride offsets
     riscvStrideOffset  offsetsLMULx4[VBYTES_MAX*4]; // LMULx4 stride offsets
     riscvStrideOffset  offsetsLMULx8[VBYTES_MAX*8]; // LMULx8 stride offsets